project/requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
passlib[bcrypt]
sqlalchemy
python-jose 
passlib
redis
email-validator